refactor(2): remove collection runtime and normalize plugin-only flow
Checkpoint removal of collection manager runtime and command surface in
favor of plugin-only operations.

- delete collection commands and collection manager app modules
- remove collection formatter wiring from runtime formatter registry
- keep plugin command surface focused on bundled and installed plugins
- normalize plugin validate messages to bundled + installed semantics
- remove cm-utils dependency from plugin create/installer paths
- keep archetype flow path-based without collection assumptions
